emoji dictionary and dictionary format
added an emoji dictionary but it is in the wrong format so there is some code to run it as a dataframe to a csv.

The code is located at the bottom of the page, the top section of code may let the csv read the unicode and convert it to emojis, while the bottom code will convert the unicode to a string to make the csv leave it as just the unicode. You may comment out whatever section of code you would not like to use